Extend general contribution validate to check financial items
